Use `trybuild` to test errors generated by gc-arena-derive
The soundness of the code generated by `gc-arena-derive` relies on
certain things *not* compiling. Currently, this is not testing, making
it possible to accidentally introduce unsoundess.

This commit uses the `trybuild` crate to add compile-fail tests. To
start out, I've added a single test that verifies that
`#[collect(no_drop)]` causes a compilation error if a `Drop` impl is
present.