Add more useful Debug, Pointer, and Display impls
This adds more useful Debug and Pointer impls for the Gc and GcCell
types, and adds a Display impl for the Gc type.  These impls mirror the
implementations for Rc and RefCell from the standard library.

Note: this does mean that you will no longer be able to derive Debug on
types that contain a Gc if they themselves contain types that don't
implement Debug.  You still can manually implement Debug in this case
however.