Decouple and expose the `GcRefCell` type publicly
This decouples the `GcRefCell` type from the `Gc` and `GcCell` types.
This simplifies things quite a bit: the `GcCell` and `GcWeakCell` types
can now just be written as type aliases.  This also paves the way for
modeling types like `Gc<'gc, Vec<GcRefCell<'gc, i32>>>`, where the
vector allocation is immutable, but individual entries in it are mutable
(and can be borrowed individually).  Users can't currently construct
such a type, but it should now be possible to expose the functionality
to do so with a little additional effort.  This comes at the cost of an
additional pointer for each `GcRefCell` allocation, which seems
reasonable.

This builds on the work done in #29.