Minor improvement: make `DynamicRootSet` Copy and use `GcCell` internally
So this is a pretty minor thing, but this improves `DynamicRootSet`
by using a `GcCell` instead of an internal `RefCell`. After trying
to use this in `luster` the first thing I ran into was wanting to
put `DynamicRootSet` in a `Gc`, but since it required &mut that meant
putting it in a `GcCell`. However, this is effectively two nested layers
of `RefCell`. It is *almost* exactly the same cost to use a `GcCell`
instead of a `RefCell` internally to `DynamicRootSet`, and this also
allows `DynamicRootSet` to implement `Clone` and `Copy`, and it can have
its own write barrier that is not shared with the root write barrier.

As a drive by improvement, also adds unsafe `GcCell::borrow_mut` and
`GcCell::try_borrow_mut`.