Fix the non-uniqueness of the `DynamicRootSet` `id` address.
This approach is a bit more intricate than it probably should be, but
it is written to both optimize the size of a `DynamicRoot` and minimize
the cost of `DynamicRootSet::fetch`. We store an `Rc<SetId>` to uniquely
identify each `DynamicRootSet`, but we also stash a clone of this inside
the `rc` field of each `DynamicRoot` handle. This is the best thing I
could think of that does not involve atomics, but it might be that a
simple static atomic counter is better.