Add `gc_arena::reroot!` macro to allow accessing `GcCell` mutably
Using this macro provides mutable access to the contents of a `GcCell`,
but with a fresh `'a` lifetime distinct from the original `'gc`
lifetime. This prevents moving or allocating pointers with the `'gc`
lifetime in or out of the `GcCell` contents. As a result,
previously-unreachable data cannot become reachable as a result
of this macro, making it sound to avoid a write barrier.

One use case for this macro is setter methods from non-GC data,
which will no longer need to take in a MutationContext.