Expose the live GC flag as `GcWeak(Cell)::is_dropped`
This is mainly intended for implementors of weak data-structures,
to allow the cleaning of stale references in `Collect::trace`, where
fully upgrading a `GcWeak` isn't possible.

An alternative API would be to expose upgrade methods on
`CollectionContext` instead, but this eems much more constraining on
the possible space of GC algorithms.