Drop the `gc-sequence` crate.
This has always been a bit of a strange API, and very very hard to use.

The new APIs for transforming Root types in `gc-arena` and for dynamic
roots would probably serve *most* use cases of this crate. Those APIs
also make the `make_sequencable_arena!` macro unnecessary, as well as
having any sort of special "sequencable" arena type, since the root
object can easily be changed to contain an in-progress sequence now,
making re-implementing the functionality of this crate much easier.

For uses that genuinely need something like `Sequence` (like maybe
`luster` if I pick it up again), it's probably better to keep a
specialized version of that API (which is what I'm going to do if I pick
up `luster` again).