Completely get rid of GcCell and GcWeakCell
They are not necessary anymore, because it is very straightforward
to just provide safe methods on the `Gc<CollectCell>` and
`Gc<CollectRefCell>` types directly.

This results in slightly more verbose code, but it is vastly simpler,
and if we can have projected `Gc` pointer types, this or something
similar to this should enable fairly ergonomic fine grained mutability.