Make mutate methods have a more restrictive signature
...that prevents obtaining a `Gc<'static>` by leaking the arena.

Allowing this unfortunately breaks the safety of making the arena Send,
because you can get a Send `Gc<'static, T>` for some T that is !Sync.

This change fixes this soundness hole, but double unfortunately breaks
the lifetime trick Ruffle uses to not have to carry three lifetimes
around everywhere in its context type.

Still investigating how to resolve these two things in a satisfactory
way...