Modify names and lifetime signatures for ergonomics.
Renames CollectionContext -> Collection and MutationContext -> Mutation.

Gets rid of the explicit extra 'a lifetime on Mutation in favor of
an easier to deal with ``&Mutation<'gc>``, and the same change to
Collection, the Collection parameter becomes `&Collection`.

Mutation and Collection are now `#[repr(transparent)]` newtypes of the
Context, and they are produced through transmute.

The methods on Arena require a callback that accepts `&'gc
MutationContext<'gc>`. This does not give any extra powers per se
and is not dangerous, but it *does* enable wrapper types around `&'gc
MutationContext<'gc>` to have one less lifetime in play.