Makes `Arena::remembered_size()` return something vaguely reasonable.
This really doesn't even need to be accessible at *all*, but currently
it mostly returns zero or small, meaningless values. The value being
returned is a *counter* that is only valid during the sweep phase, and
there is no way to even tell what phase the collector is in from the
outside.

Instead of doing that, once we are finished with the sweep phase, keep
track of the total remembered size for that pass, and always return the
remembered size from the *last* sweep phase. This way, the value is at
least a decent estimate of whatever the current remembered size would
be.