Collect for allocator-api2 and hashbrown, also Metrics allocator
Implements `Collect` for HashMap and HashSet from hashbrown, also Vec
and Box from allocator-api2. Also implements a custom `Allocator` trait
wrapper that records allocations and deallocations from gc-arena's
`Metrics` handle, and is branded with the branding lifetime so it cannot
be mixed up with other arena types.

The implementation for hashbrown is slightly subtle because it is
different depending on whether or not the "allocator-api2" feature is
also enabled. Circle-CI now also checks all combinations of the features
"allocator-api2" and "hashbrown" to make sure they all compile.