Different strategy for counting DynamicRootSet sizes
We no longer use the `Metrics` handle as the set ID, because it is
important that we distinguish one set from another. Otherwise this could
lead to unsafety, where a single set is dropped but another is not, a
handle would be accessible without actually being traced.

Additionally, we store the size of the held T parallel with the
weak pointer, and we only decrement the external allocation when the
*both* the Rc and Weak are dropped, this is more accurate (because the
store lives until the last of Rc or Weak are dropped) and also better
encourages the root list to be collected when a lot of external handles
are created then dropped.