Finalization support part 3
Makes as few decisions as possible and leaves everything up to the user.

Instead of having *any* direct finalization support, "simply" allows
the user to stop collection at exactly the end of the marking phase and
examine things.

Adds a `finalize` (indirect) method on `Arena`, and a new `Finalization`
context which is currently `Mutation` but with two new powers:

* The ability to check whether a `GcWeak` was "dead" at the start of the
  current finalization call
* The ability to "resurrect" a `GcWeak` to preserve it this cycle (even
  without being upgraded and stored).

Entirely punts on ordering and queueing and leaves everything up to the
user, BUT it is just powerful enough to implement finalizers and Java
style ReferenceQueues and ephemeron tables and finalization ordering and
many other things that parts 1 or 2 (or both) struggled with.