Add GcWeak::trace_strong
Allows someone to more easily decide at runtime whether a GcWeak should
be treated as a strong pointer or a weak pointer.

This is not *strictly* required for `piccolo` to implement weak tables
but it makes it a lot less obnoxious.

Includes a couple of drive by changes adding extra debug asserts, and a
*very small* perf increase for GcWeak::resurrect.