Make GcWeak::is_dropped only valid during collection
During mutation, GcWeak::upgrade is a strictly better check. During
collection, we know that we are *not* in the sweep phase, so we don't
have to document the confusing property that `GcWeak::is_dropped` and
`GcWeak::upgrade` can be mismatched.

Includes a couple of drive by changes adding extra debug asserts, and a
*very small* perf increase for GcWeak::resurrect.