Simplest possible forward write barriers.
This has some additional drive-by cleanup to make things more
consistent. It adds two methods to `Mutation`, one for backwards write
barriers (previously the only kind) and one for forward write barriers.

Both backwards and forward write barriers are changed to take one or
optionally two pointers: the parent and the adopted child. If both
pointers are given, then the barriers check if the parent is black and
the child is white, and if so, invoke the barrier. The backwards write
barrier will turn the parent gray, and the forwards write barrier will
turn the child gray.

If the backwards write barrier is only given the parent pointer, then it
unconditionally turns it gray if it was black, allowing it to adopt any
pointers at all. If the forwards write barrier is only given the child
pointer, then it unconditionally turns it gray if it was white, allowing
it to be adopted by any possible pointer.