Provide a way of creating a 'static MetricsAlloc
It has no actual 'gc references, the 'gc lifetime is purely a lint,
so this is fine to provide. It is never a safety issue to misuse the
metrics object, and this was already possible anyway (it's not that
difficult to reimplement `MetricsAlloc` safely).

Sometimes having a non-'static allocator is *incredibly* annoying, so
this is useful.