Simplify forward_barrier
Don't duplicate the check child is white check with Context::trace and
also outline it, as this could be a pessimization. Wait for actual
evidence before doing anything other than the simplest thing.

Also add a note about a harmless corner case with the backward barrier
being called on a pointer where `Collect::needs_trace()` is false