Add 'Arena::get_mut_gc' api to allow mutating a `Gc`
This method takes `&mut self`, so it can only be used
outside of a `mutate`/`mutate_root` call.

This is useful when an application (e.g. Ruffle) almost always
read from a particular `Gc`, and occasionally need to write. If
modifying the `Gc` can be done outside of a `mutate` call,
then this API can be used to avoid introducing a `GcCell`