Move `TraceExt::trace` to an inherent method on `Trace`
Uses a slightly different design that does not require that
`Collect::trace` accept a `?Sized` impl of `Trace` (which is fine
because `Collect` is never object safe anyway).

Since methods that actually *use* `T: Trace` impls require that `T:
Sized`, we can make the extension method inherent on the `Trace` trait.

This change keeps `Trace` object safe, and we may use this in the future
to implement tracing through opaque `Gc` pointers. Since this will need
to be implemented through the custom CollectVtable, needing to make a
custom wrapper around a `&mut dyn Trace` is not a big deal, and this is
the ONLY need to ever have a `&mut dyn Trace` in the first place.