Improvements and further simplifications to the new pacing code
1. Don't have a `CycleResult` and get rid of `Arena::collect_all` The
old concept of a "full" collection was not that useful because it didn't
take into account the situation where there was any sort pointer marked
as reachable that is logically "forgotten" during the cycle (as in,
through mutation but after the pointer is marked as reachable). It is
too confusing to explain to the user what this means, and it is arguable
that it isn't even useful, so don't include it.

2. Only reset accumulated debt when an ENTIRE cycle is performed as an
atomic unit. This still keeps accumulated debt from spiraling out of
control in cases where there is a large amount of induced debt or when
the arena is very small and min_sleep is small and every collect_debt is
a full collection, but it keeps debt in any *possible* case where
collecting more might *conceivably* reduce memory.

3. Improve the documentation for `Pacing` and describe some of the edge
case behaviors like elided work and weakly reachable values.

4. Move the `Pacing` constants into `Pacing`, and just remove
`Pacing::stop_the_world()` in favor of `Pacing::STOP_THE_WORLD`.

5. Change what a "mark" is slightly to include marking a value as weakly
reachable. Now, a value is "marked" when it is first transitioned from
white to any other color, and only the first time this happens. This
means that both strongly and weakly reachable pointers have a "mark",
and this is documented in the more complete `Pacing` docs.