Add `mark_gc_untraced` to `Metrics` and use it.
With this change, placing a pointer into the `gray_again` queue *undoes*
the work that was performed when originally tracing it.

This makes backwards write barriers actually effectively incur some
*debt* that will be paid back during processing of the `gray_again`
queue.

This is a strict improvement from the situation before. Before, the
entire `gray_again` queue performed *no* work (so that work was not
double counted), but this meant that there could be a huge pause time
when processing it. Now, the incurred *debt* from the write barrier
moving the pointer back into the gray queue can be processed
immediately. In the situation before, this "debt" that the write
barriers incurred was not directly recorded, it only was represented
by the size of the `gray_again` queue, and was "paid off" during a
huge atomic step at the end of tracing. This change makes this debt
"official", and results in better pacing and lower pause time.