Add `Arena::cycle_debt` and some method renames to aid phase tracking
Since `Arena::collect_debt` no longer deterministically returns at exact
end of a cycle (because debt is now normally carried over to the next
cycle), there is no way to always be able to track when a new collection
cycle starts.

This fixes the problem by adding `Arena::cycle_debt`, which works like
the old `Arena::collect_debt` in that it always returns at
the end of a cycle. It also renames `Arena::finish_collection`
to `Arena::finish_cycle` to be in symmetry the same way as
`Arena::mark_debt` and `Arena::finish_marking`, and improves the
comments to make the situation easier to understand.